got ride of some warnings and added some stuff

+++ FallRiver/FallRiver/GamePlayState.cpp
@@ -455,7 +455,7 @@ void GamePlayState::Enter()
 		}
 		else if( _stricmp(nth->m_cType,"Town") == 0)
 		{
-			m_pHUD->SetTarget(nth->x, nth->y);
+			m_pHUD->SetTarget(float(nth->x), float(nth->y));
 		}
 		else if ( _stricmp(nth->m_cType,"Spawn Point") == 0 )
 		{
@@ -1467,7 +1467,7 @@ void GamePlayState::HandleEvent(Event* aPEvent)
 {
 	if(aPEvent->GetEventID() == "ForestToTown")///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 	{
-		for( int i = 0; i < this->fireA.size(); i++)
+		for( unsigned int i = 0; i < this->fireA.size(); i++)
 		{
 			Particle_Manager::GetInstance()->GetActiveEmitter(fireA[i])->SetLoopin(false);
 		}
@@ -1475,7 +1475,7 @@ void GamePlayState::HandleEvent(Event* aPEvent)
 	
 	if(aPEvent->GetEventID() == "HouseToTown")///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 	{
-		for( int i = 0; i < this->fireA.size(); i++)
+		for( unsigned int i = 0; i < this->fireA.size(); i++)
 		{
 			Particle_Manager::GetInstance()->GetActiveEmitter(fireA[i])->SetLoopin(false);
 		}
@@ -1484,7 +1484,7 @@ void GamePlayState::HandleEvent(Event* aPEvent)
 	
 	if(aPEvent->GetEventID() == "HospitalToTown")///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 	{
-		for( int i = 0; i < this->fireA.size(); i++)
+		for( unsigned int i = 0; i < this->fireA.size(); i++)
 		{
 			Particle_Manager::GetInstance()->GetActiveEmitter(fireA[i])->SetLoopin(false);
 		}
@@ -1492,7 +1492,7 @@ void GamePlayState::HandleEvent(Event* aPEvent)
 	
 	if(aPEvent->GetEventID() == "GoToHospital")///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 	{
-		for( int i = 0; i < this->fireA.size(); i++)
+		for( unsigned int i = 0; i < this->fireA.size(); i++)
 		{
 			Particle_Manager::GetInstance()->GetActiveEmitter(fireA[i])->SetLoopin(false);
 		}
@@ -1501,7 +1501,7 @@ void GamePlayState::HandleEvent(Event* aPEvent)
 	
 	if(aPEvent->GetEventID() == "GoToHouse")///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 	{
-		for( int i = 0; i < this->fireA.size(); i++)
+		for( unsigned int i = 0; i < this->fireA.size(); i++)
 		{
 			Particle_Manager::GetInstance()->GetActiveEmitter(fireA[i])->SetLoopin(false);
 		}
@@ -1707,7 +1707,7 @@ void GamePlayState::HandleEvent(Event* aPEvent)
 			}
 			else if( _stricmp(nth->m_cType,"Town") == 0)
 			{
-				m_pHUD->SetTarget(nth->x, nth->y);
+				m_pHUD->SetTarget((float)nth->x, (float)nth->y);
 			}
 			else if( _stricmp(nth->m_cType,"Hospital") == 0 )
 			{
@@ -2037,7 +2037,7 @@ void GamePlayState::HandleEvent(Event* aPEvent)
 				}
 				else if( pPlayer->m_bHasMedicine == false  && pPlayer->m_vpActiveQuests.size() > 0 )
 				{
-					m_pHUD->SetTarget(hospitalX, hospitalY);
+					m_pHUD->SetTarget((float)hospitalX, (float)hospitalY);
 				}
 				else if( pPlayer->m_bHasMedicine == false && pPlayer->m_vpActiveQuests.size() == 0 )
 				{
@@ -2051,7 +2051,7 @@ void GamePlayState::HandleEvent(Event* aPEvent)
 			}
 			else if( _stricmp(nth->m_cType,"Town") == 0)
 			{
-				m_pHUD->SetTarget(nth->x, nth->y);
+				m_pHUD->SetTarget((float)nth->x, (float)nth->y);
 			}
 			else if( _stricmp(nth->m_cType,"Hospital") == 0 )
 			{
